feat: 역 추가/삭제 기능 구현

+++ src/main/java/subway/domain/Path.java
@@ -3,9 +3,9 @@ package subway.domain;
 public final class Path {
 
     private final Station next;
-    private final int distance;
+    private final Long distance;
 
-    public Path(final Station next, final int distance) {
+    public Path(final Station next, final Long distance) {
         this.next = next;
         this.distance = distance;
     }
@@ -14,7 +14,7 @@ public final class Path {
         return next;
     }
 
-    public int getDistance() {
+    public Long getDistance() {
         return distance;
     }
 }
